requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
MarkupSafe==2.1.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
Werkzeug==3.0.1
